@@ -2,20 +2,24 @@
   "name": "angular-samples",
   "scripts": {
     "build-all": "bazel build //src/apps/api && bazel build //src/apps/backend",
-    "start": "concurrently \"ibazel run //src/apps/api\" \"ibazel run //src/apps/backend\""
+    "start": "concurrently \"ibazel run //src/apps/api\" \"ibazel run //src/apps/backend\"",
+    "postinstall": "ngc -p angular-metadata.tsconfig.json"
   },
   "devDependencies": {
-    "@angular/bazel": "7.2.2",
-    "@angular/compiler": "7.2.2",
-    "@angular/compiler-cli": "7.2.2",
-    "@bazel/ibazel": "0.9.1",
-    "@bazel/typescript": "0.20.0",
+    "@angular/bazel": "8.0.0-rc.0",
+    "@angular/cli": "8.0.0-beta.10",
+    "@angular/compiler": "8.0.0-rc.0",
+    "@angular/compiler-cli": "8.0.0-rc.0",
+    "@bazel/buildifier": "0.22.0",
+    "@bazel/ibazel": "0.10.1",
+    "@bazel/typescript": "0.27.12",
+    "@bazel/karma": "0.27.12",
     "@types/node": "^10.12.0",
     "@types/react": "16.4.1",
     "compression": "1.7.2",
     "concurrently": "4.1.0",
     "reload": "2.3.1",
-    "typescript": "3.1.1",
+    "typescript": "3.4.5",
     "webpack": "4.20.2",
     "webpack-cli": "3.1.2"
   },
@@ -28,6 +32,15 @@
     "react": "16.4.2",
     "react-dom": "16.4.2",
     "request": "2.87.0",
-    "zone.js": "0.8.26"
+    "zone.js": "0.9.0",
+    "@angular/common": "8.0.0-rc.0",
+    "@angular/core": "8.0.0-rc.0",
+    "@angular/forms": "8.0.0-rc.0",
+    "@angular/platform-browser": "8.0.0-rc.0",
+    "@angular/platform-browser-dynamic": "8.0.0-rc.0",
+    "@angular/router": "8.0.0-rc.0",
+    "@angular/upgrade": "8.0.0-rc.0",
+    "@angular/animations": "8.0.0-rc.0",
+    "rxjs": "6.4.0"
   }
 }
