@@ -1,14 +1,14 @@
 import {NgModule} from '@angular/core';
 import {RouterModule} from '@angular/router';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {HttpModule} from '@angular/http';
+import {HttpClientModule} from '@angular/common/http';
 import {CommonModule} from '@angular/common';
 
 @NgModule({
   exports: [
     FormsModule,
     ReactiveFormsModule,
-    HttpModule,
+    HttpClientModule,
     CommonModule,
     RouterModule,
   ],
