@@ -1,17 +1,13 @@
 import {Routes, RouterModule} from '@angular/router';
 import {ModuleWithProviders} from '@angular/core';
-
 import {DemoPage} from '../shared-components/demo-page';
 import {AboutComponent} from '../shared-components/about/about.component';
-
 import {TreeViewDemo} from '../shared-components/tree-view/tree-view-demo';
 import {ContactList} from '../shared-components/contact-list/contact-list';
 import {BoundTextbox} from '../shared-components/bound-textbox/bound-textbox';
 import {GridDemo} from '../shared-components/grid/grid-demo';
-import {HttpSample} from '../shared-components/http/http';
 import {Spreadsheet} from '../shared-components/spreadsheet/spreadsheet';
 import {Algorithms} from '../shared-components/algorithms/algorithms';
-import {AngularHost} from '../shared-components/react-integration/angular-host';
 import {JqueryIntegration} from '../shared-components/jquery-integration/jquery-integration';
 import {InputControls} from '../shared-components/input-controls/input-controls';
 import {AddressBook} from '../shared-components/dependency-injection/address-book';
@@ -31,7 +27,7 @@ import {RxJsBuffering} from '../shared-components/rxjs-buffering/rxjs-buffering'
 import {CarsComponent} from '../shared-components/rxjs-error-handling/cars.component';
 import {SvelteComponent} from '../shared-components/svelte-integration/svelte.component';
 
-const routes: Routes = [
+export const routes: Routes = [
   {
     path: '',
     redirectTo: '/demo/spreadsheet',
@@ -44,13 +40,11 @@ const routes: Routes = [
     children: [
       {path: 'spreadsheet', component: Spreadsheet},
       {path: 'jquery', component: JqueryIntegration},
-      {path: 'react', component: AngularHost},
       {path: 'algorithms', component: Algorithms},
       {path: 'graph', component: Graph},
       {path: 'pub-sub', component: PubSub},
       {path: 'text-editor', component: TextEditor},
       {path: 'address', component: AddressBook},
-      {path: 'http', component: HttpSample},
       {path: 'treeview', component: TreeViewDemo},
       {path: 'grid', component: GridDemo},
       {path: 'input', component: InputControls},
@@ -71,4 +65,3 @@ const routes: Routes = [
     ],
   },
 ];
-export const routing: ModuleWithProviders = RouterModule.forRoot(routes);
