@@ -1,4 +1,4 @@
-import {Http} from '@angular/http';
+import {HttpClient} from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {IPerson} from '../../models';
 import {map} from 'rxjs/operators';
@@ -6,16 +6,13 @@ import {Observable} from 'rxjs';
 
 @Injectable()
 export class AddressBookService {
-  http: Http;
-  constructor(http: Http) {
+  http: HttpClient;
+  constructor(http: HttpClient) {
     console.log('Creating AddressBookService');
     this.http = http;
   }
 
   getEntries(): Observable<Array<IPerson>> {
-    return this.http.get('/api/people').pipe(
-      map(res => res.json()),
-      map((res: Array<IPerson>) => res),
-    );
+    return this.http.get('/api/people').pipe(map((res: Array<IPerson>) => res));
   }
 }
