@@ -1,14 +1,14 @@
 import {Injectable} from '@angular/core';
-import {Http, Response} from '@angular/http';
+import {HttpClient} from '@angular/common/http';
 import {Observable} from 'rxjs';
-
+import {IFriends} from '../../models';
 import {publishReplay, map, tap, refCount} from 'rxjs/operators';
 
 @Injectable()
 export class FriendsService {
   _friends: Observable<any> = null;
 
-  constructor(private _http: Http) {}
+  constructor(private _http: HttpClient) {}
 
   clearCache() {
     this._friends = null;
@@ -17,8 +17,7 @@ export class FriendsService {
   getFriends() {
     if (!this._friends) {
       this._friends = this._http.get('/api/friends').pipe(
-        map((res: Response) => res.json().friends),
-        tap(friends => console.log('fetched friends')),
+        map((res: IFriends) => res.friends),
         publishReplay(1),
         refCount(),
       );
